pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "forgetit"
version = "1.0.0a0"
readme = "README.md"
requires-python = ">=3.12"
authors = [{ name = "Miguel Arroyo" }]

dependencies = [
    "falkordb",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0",
    "humanize"
]

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]
include = ["forgetit*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
